+++ samples/Game.Unity.Agar/Server/Hotfix/State/Rooms/RoomBehavior.cs
@@ -2,6 +2,7 @@ using Server.App.State.Contracts;
 using Server.App.State.Contracts.Leaderboard;
 using Server.App.State.Contracts.Rooms;
 using Server.App.State.Contracts.Sessions;
+using Server.App.State.Contracts.Timers;
 using Server.App.State.Contracts.Users;
 using Server.App.State.Leaderboard;
 using Server.App.State.Rooms;