+++ samples/Game.Unity.Agar/Server/Hotfix/Timers/BattleRuntimeTimerCallbacks.cs
@@ -1,5 +1,6 @@
 using Server.App.State.Contracts;
 using Server.App.State.Contracts.Rooms;
+using Server.App.State.Contracts.Timers;
 using Server.App.State.Rooms;
 using Lakona.Game.Server.Actors;
 using Lakona.Game.Server.Hotfix;