+++ samples/Game.Unity.Agar/Server/Hotfix/Timers/MatchmakingTimerArgs.cs
@@ -1,6 +1,0 @@
-namespace Server.Hotfix.Timers;
-
-public sealed class MatchmakingTimerArgs
-{
-    public string OwnerActorId { get; init; } = "";
-}